requirements.txt
langchain==0.1.5
langchain-openai==0.0.5
python-dotenv==1.0.0
openai==1.12.0
pydantic==2.5.2
pytest==7.4.3
streamlit==1.32.0 